Follow Sistemi eklendi

+++ src/main/java/tr/anil/questapp/dao/FollowDao.java
@@ -16,8 +16,8 @@ public interface FollowDao extends JpaRepository<Follow, Long> {
     Follow findByFollowedUserIdAndFollowerUserId(@Param("followedId") Long followedId, @Param("followerId") Long followerId);
 
     @Query(value = "select count(*) from _follow where followed_user_id = :userId", nativeQuery = true)
-    Long getFollowedCountByUserId(@Param("userId") Long userId);
+    int getFollowedCountByUserId(@Param("userId") Long userId);
 
     @Query(value = "select count(*) from _follow where follower_user_id = :userId", nativeQuery = true)
-    Long getFollowerCountByUserId(@Param("userId") Long userId);
+    int getFollowerCountByUserId(@Param("userId") Long userId);
 }
